Fix applying subsettings, fix LimitscolumnWidget
