+++ openc3-cosmos-init/plugins/packages/openc3-cosmos-demo/targets/INST2/screens/limits.txt
@@ -6,28 +6,33 @@ HORIZONTAL
   VERTICALBOX "Limits Bars"
     LIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
     VALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
+      SUBSETTING 0 WIDTH 130
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
     # The TEMP1 limits are defined as follow:
     # LIMITS DEFAULT 1 ENABLED -80.0 -70.0 60.0 80.0 -20.0 20.0
     # Thus the following are examples of limiting the limitsbar
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -80
-      SUBSETTING 1 MAX_VALUE 80
+      SUBSETTING 1 WIDTH 130
+      SUBSETTING 2 MIN_VALUE -80
+      SUBSETTING 2 MAX_VALUE 80
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -75
-      SUBSETTING 1 MAX_VALUE 70
+      SUBSETTING 1 WIDTH 130
+      SUBSETTING 2 MIN_VALUE -75
+      SUBSETTING 2 MAX_VALUE 70
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -70
-      SUBSETTING 1 MAX_VALUE 60
+      SUBSETTING 1 WIDTH 130
+      SUBSETTING 2 MIN_VALUE -70
+      SUBSETTING 2 MAX_VALUE 60
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -50
-      SUBSETTING 1 MAX_VALUE 50
+      SUBSETTING 1 WIDTH 130
+      SUBSETTING 2 MIN_VALUE -50
+      SUBSETTING 2 MAX_VALUE 50
     LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -20
-      SUBSETTING 1 MAX_VALUE 20
-    LABELVALUELIMITSBAR <%= target_name %> HEALTH_STATUS TEMP1
-      SUBSETTING 1 MIN_VALUE -10
-      SUBSETTING 1 MAX_VALUE 10
+      SUBSETTING 0 TEXTCOLOR grey
+      SUBSETTING 1 WIDTH 130
+      SUBSETTING 1 BORDERCOLOR grey
+      SUBSETTING 2 MIN_VALUE -20
+      SUBSETTING 2 MAX_VALUE 20
   END
   VERTICAL
     VERTICALBOX "Range Bars 0-100"
@@ -46,21 +51,22 @@ END
 HORIZONTAL
   VERTICALBOX "Limits Columns"
     HORIZONTAL
-      LIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2
-      VALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2
-      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2 WITH_UNITS 8
-      # The TEMP2 limits are defined as follow:
-      # LIMITS DEFAULT 1 ENABLED -60.0 -55.0 30.0 35.0
+      LIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP3
+        SETTING HEIGHT 200
+      VALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP3
+        SUBSETTING 0 WIDTH 130
+      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP3
+        SUBSETTING 1 WIDTH 130
+      # The TEMP3 limits are defined as follow: RL/-25 YL/-10 YH/50 RH/55
       # Thus the following are examples of limiting the limitscolumn
-      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2 WITH_UNITS 8
-        SUBSETTING 1 MIN_VALUE -60
-        SUBSETTING 1 MAX_VALUE 35
-      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2 WITH_UNITS 8
-        SUBSETTING 1 MIN_VALUE -55
-        SUBSETTING 1 MAX_VALUE 30
-      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP2 WITH_UNITS 8
-        SUBSETTING 1 MIN_VALUE -20
-        SUBSETTING 1 MAX_VALUE 10
+      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP3
+        SUBSETTING 1 WIDTH 130
+        SUBSETTING 2 MIN_VALUE -25
+        SUBSETTING 2 MAX_VALUE 55
+      LABELVALUELIMITSCOLUMN <%= target_name %> HEALTH_STATUS TEMP3
+        SUBSETTING 1 WIDTH 130
+        SUBSETTING 2 MIN_VALUE -10
+        SUBSETTING 2 MAX_VALUE 50
     END
   END
